Fixed package name, added support for different media types, improved documentation.
