+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Change Log
+
+## 0.1.1
+
+- Improved documentation.
+- Support for multiple file types added.
+- Changed package name when I decided to buy `lubecki.com` but the minimum bid was 5000£. Managed to get `jlubecki.com`.
