+++ src/main/java/com/jlubecki/q/Q.java
@@ -22,14 +22,14 @@
  * SOFTWARE.
  */
 
-package com.lubecki.q;
+package com.jlubecki.q;
 
 
-import com.lubecki.q.logging.LogLevel;
-import com.lubecki.q.logging.QLog;
-import com.lubecki.q.playback.Player;
-import com.lubecki.q.playback.PlayerManager;
-import com.lubecki.q.playback.PlayerState;
+import com.jlubecki.q.playback.PlayerState;
+import com.jlubecki.q.logging.LogLevel;
+import com.jlubecki.q.logging.QLog;
+import com.jlubecki.q.playback.Player;
+import com.jlubecki.q.playback.PlayerManager;
 import org.intellij.lang.annotations.RegExp;
 import org.jetbrains.annotations.NonNls;
 
@@ -56,7 +56,6 @@ public final class Q {
 
     private static Q instance;
 
-
     //region Error Messages
 
     private static final String UNKNOWN_LOOP_STATE = "Unknown loop state.";
@@ -64,10 +63,7 @@ public final class Q {
 
     //endregion
 
-
     //region Logging
-
-    private static final String LOG_TAG = "Q";
 
     /**
      * Returns the singleton instance of the Q object. The lazy initialization of the singleton object is synchronized
@@ -88,11 +84,11 @@ public final class Q {
      * @param prefix Simple modifier to help identify what method changed the index / where the index is being logged.
      */
     private void logIndex(@NonNls String prefix) {
-        if (QLog.getLogLevel() != LogLevel.PLAYER) {
+        if (QLog.getLogLevel() != LogLevel.NONE && QLog.getLogLevel() != LogLevel.PLAYER) {
             String message = prefix + "  ::  Index: %d    Track: %s";
             message = String.format(message, index, current.title);
 
-            QLog.i(LOG_TAG, message);
+            QLog.i(this.getClass().getSimpleName(), message);
         }
     }
 
@@ -116,7 +112,7 @@ public final class Q {
                 message = eventName + "  ::  Track: %s";
                 message = String.format(message, current.title);
 
-                QLog.v(LOG_TAG, message);
+                QLog.v(this.getClass().getSimpleName(), message);
                 break;
 
             case Q:
@@ -126,7 +122,7 @@ public final class Q {
                 message = String.format(message, current.title, current.artist, current.uri, current.image, state,
                         index);
 
-                QLog.v(LOG_TAG, message);
+                QLog.v(this.getClass().getSimpleName(), message);
                 break;
         }
     }
@@ -135,15 +131,22 @@ public final class Q {
      * Logs state changes. Log priority is VERBOSE.
      */
     private void logState() {
-        if (QLog.getLogLevel() == LogLevel.FULL && QLog.getLogLevel() == LogLevel.Q) {
+        if (QLog.getLogLevel() == LogLevel.FULL || QLog.getLogLevel() == LogLevel.Q) {
             String message = String.format("Q State changed to: %s", state.toString());
 
-            QLog.v(LOG_TAG, message);
+            QLog.v(this.getClass().getSimpleName(), message);
+        }
+    }
+
+    private void logMediaType() {
+        if (QLog.getLogLevel() == LogLevel.FULL || QLog.getLogLevel() == LogLevel.Q) {
+            String message = String.format("Current media type changed to: %s", current.mediaType.toString());
+
+            QLog.v(this.getClass().getSimpleName(), message);
         }
     }
 
     //endregion
-
 
     //region Playback
 
@@ -156,21 +159,46 @@ public final class Q {
     private QTrack next;
     private Loop loop = Loop.NONE;
     private QState state;
+    private MediaType mediaType;
     private boolean resetOnPrevious = true;
     private int minDelay = 2000;
 
+    //region Public Methods
+
+    /**
+     * Adds a player to the {@link PlayerManager}.
+     *
+     * @param uriPattern the pattern linked to the provided {@link Player}.
+     * @param player     the {@link Player} that should handle a track uri matching the provided pattern.
+     */
     public void addPlayer(@RegExp String uriPattern, Player player) {
         manager.registerPlayer(uriPattern, player);
     }
 
+    /**
+     * Removes a player from the {@link PlayerManager}.
+     *
+     * @param uriPattern the pattern linked to the {@link Player} that should be removed.
+     */
     public void removePlayer(@RegExp String uriPattern) {
         manager.unregisterPlayer(uriPattern);
     }
 
+    /**
+     * Recommended to get callbacks for the state of the Q. Some use cases may not need to listen for {@link QState}
+     * changes.
+     *
+     * @param listener the object that should listen for changes to the state of the Q.
+     */
     public void setListener(QEventListener listener) {
         this.listener = listener;
     }
 
+    /**
+     * Sets the list of tracks for the Q to play.
+     *
+     * @param tracks the list of tracks that should be played.
+     */
     public void setTrackList(List<? extends QTrack> tracks) {
         this.tracks = new ArrayList<QTrack>(tracks);
         this.original = new ArrayList<QTrack>(tracks);
@@ -182,76 +210,8 @@ public final class Q {
     }
 
     /**
-     * Increases the index. Loops to the beginning if already at the end of the list.
+     * Moves to the previous track.
      */
-    private void increment() {
-        index++;
-        index = index > tracks.size() - 1 ? 0 : index;
-        setCurrent();
-
-        logIndex("INCREMENT");
-    }
-
-    /**
-     * Decreases the index. Loops to the end of the list if already at index 0.
-     */
-    private void decrement() {
-        index--;
-        index = index < 0 ? tracks.size() - 1 : index;
-        setCurrent();
-
-        logIndex("DECREMENT");
-    }
-
-    /**
-     * Simply sets the current track. Kills the Q immediately if a user manages to try to play something when nothing is
-     * there to be played.
-     *
-     * @throws IllegalStateException when the user tries to play an empty list.
-     */
-    private void setCurrent() {
-        if (validIndex()) {
-            current = tracks.get(index);
-            manager.updatePlayer(current);
-        } else {
-            String message = tracks != null ?
-                    String.format("Couldn't set current track  ::  Index: %d    Tracks: %d", index, tracks.size()) :
-                    "Track list was null.";
-
-            IllegalStateException exception = new IllegalStateException(message);
-
-            if (QLog.shouldIgnoreIllegalStates()) {
-                QLog.e(exception, this.getClass().getSimpleName(), exception.getMessage());
-            } else {
-                throw exception;
-            }
-        }
-    }
-
-    /**
-     * Plays a track for a given index. This should be used when a user selects a track to play.
-     *
-     * @param index Index of the track to play.
-     * @throws IllegalStateException whenever the index is out of bounds.
-     */
-    public void setIndex(int index) {
-        this.index = index;
-
-        stop();
-        setCurrent();
-        start();
-
-        logIndex("SetIndex");
-    }
-
-    @SuppressWarnings({"OverlyComplexBooleanExpression", "BooleanMethodNameMustStartWithQuestion"})
-    private boolean validIndex() {
-        return index >= 0 &&
-                index < tracks.size() &&
-                tracks != null &&
-                !tracks.isEmpty();
-    }
-
     public void previous() {
         if (manager.getCurrentPlayer().getCurrentTime() < minDelay || !resetOnPrevious) {
             stop();
@@ -302,6 +262,24 @@ public final class Q {
                 start();
                 break;
         }
+    }
+
+    /**
+     * Plays a track for a given index. This should be used when a user selects a track to play.
+     *
+     * @param index Index of the track to play from the unshuffled list.
+     * @throws IllegalStateException whenever the index is out of bounds.
+     */
+    @SuppressWarnings("SpellCheckingInspection")
+    public void setIndex(int index) {
+        // Corrects for shuffled list
+        this.index = tracks.indexOf(original.get(index));
+
+        stop();
+        setCurrent();
+        start();
+
+        logIndex("SetIndex");
     }
 
     /**
@@ -373,6 +351,16 @@ public final class Q {
         this.loop = loop;
     }
 
+    /**
+     * Seeks to a specific point in a song.
+     *
+     * @param time The time in the song to seek to.
+     */
+    public void seekTo(int time) {
+        manager.seekTo(time);
+
+        logPlayback("SEEKING");
+    }
 
     /**
      * Determines whether or not to simply reset the current track when previous is hit, and if so, how much time must
@@ -389,21 +377,6 @@ public final class Q {
     }
 
     /**
-     * Sets the current state for the Q.
-     *
-     * @param state New state of the Q.
-     */
-    private void setState(QState state) {
-        this.state = state;
-
-        logState();
-
-        if(listener != null) {
-            listener.onEvent(state);
-        }
-    }
-
-    /**
      * Method to get the current track.
      *
      * @return the current track.
@@ -412,6 +385,9 @@ public final class Q {
         return current;
     }
 
+    /**
+     * Gets the Q ready to begin playback without starting playback.
+     */
     public void prepare() {
         setState(QState.STARTING);
 
@@ -425,33 +401,11 @@ public final class Q {
      * anything currently playing.
      */
     public void start() {
-        // TODO Stop existing playback before starting new playback.
-
         setState(QState.STARTING);
 
-        // Player code goes here.
         manager.prepare(current.uri);
 
         logPlayback("START");
-    }
-
-    /**
-     * Should stop the current player in an intelligent way. If replaying the same track, just set to beginning. If
-     * moving to the next track, release resources if possible, etc. TODO
-     */
-    public void stop() {
-        // Player code goes here.
-        manager.stop();
-
-        setState(QState.STOPPED);
-
-        logPlayback("STOP");
-    }
-
-    public void seekTo(int time) {
-        manager.seekTo(time);
-
-        logPlayback("SEEKING");
     }
 
     /**
@@ -476,7 +430,6 @@ public final class Q {
      * Pauses playback. Should only be called if playback is ongoing.
      */
     public void pause() {
-        // Player code goes here.
         manager.pause();
 
         setState(QState.PAUSED);
@@ -484,11 +437,23 @@ public final class Q {
         logPlayback("PAUSE");
     }
 
+
+    /**
+     * Should stop the current player in an intelligent way. If replaying the same track, just set to beginning. If
+     * moving to the next track, release resources if possible, etc. TODO
+     */
+    public void stop() {
+        manager.stop();
+
+        setState(QState.STOPPED);
+
+        logPlayback("STOP");
+    }
+
     /**
      * Calls release method for all player objects and makes the Q instance null.
      */
     public void release() {
-        // Player code goes here.
         manager.removeAllPlayers();
         manager.release();
         instance = null;
@@ -497,6 +462,106 @@ public final class Q {
 
         QLog.log("Q RELEASED.");
     }
+
+    //endregion
+
+    //region Private Methods
+
+    /**
+     * Increases the index. Loops to the beginning if already at the end of the list.
+     */
+    private void increment() {
+        index++;
+        index = index > tracks.size() - 1 ? 0 : index;
+        setCurrent();
+
+        logIndex("INCREMENT");
+    }
+
+    /**
+     * Decreases the index. Loops to the end of the list if already at index 0.
+     */
+    private void decrement() {
+        index--;
+        index = index < 0 ? tracks.size() - 1 : index;
+        setCurrent();
+
+        logIndex("DECREMENT");
+    }
+
+    /**
+     * Simply sets the current track. Kills the Q immediately if a user manages to try to play something when nothing is
+     * there to be played.
+     *
+     * @throws IllegalStateException when the user tries to play an empty list.
+     */
+    private void setCurrent() {
+        if (validIndex()) {
+            current = tracks.get(index);
+            manager.updatePlayer(current);
+            setMediaType(current.mediaType);
+        } else {
+            String message = tracks != null ?
+                    String.format("Couldn't set current track  ::  Index: %d    Tracks: %d", index, tracks.size()) :
+                    "Track list was null.";
+
+            IllegalStateException exception = new IllegalStateException(message);
+
+            if (QLog.shouldIgnoreIllegalStates()) {
+                QLog.e(exception, this.getClass().getSimpleName(), exception.getMessage());
+            } else {
+                throw exception;
+            }
+        }
+    }
+
+    @SuppressWarnings({"OverlyComplexBooleanExpression", "BooleanMethodNameMustStartWithQuestion"})
+    private boolean validIndex() {
+        return index >= 0 &&
+                index < tracks.size() &&
+                tracks != null &&
+                !tracks.isEmpty();
+    }
+
+
+    /**
+     * Sets the current state for the Q.
+     *
+     * @param state New state of the Q.
+     */
+    private void setState(QState state) {
+        this.state = state;
+
+        logState();
+
+        if (listener != null) {
+            listener.onEvent(state);
+        } else {
+            QLog.w(this.getClass().getSimpleName(), "QEventCallback was null.");
+        }
+    }
+
+    /**
+     * Used by {@link #setCurrent()} to fire off a {@link QEventListener#onMediaTypeChanged(MediaType)} event.
+     *
+     * @param mediaType the new media type as determined by the current track.
+     */
+    private void setMediaType(MediaType mediaType) {
+        // only make adjustments when media type actually changes
+        if(this.mediaType != mediaType) {
+            this.mediaType = mediaType;
+
+            logMediaType();
+
+            if (listener != null) {
+                listener.onMediaTypeChanged(mediaType);
+            } else {
+                QLog.w(this.getClass().getSimpleName(), "QEventCallback was null.");
+            }
+        }
+    }
+
+    //endregion
 
     //endregion
 }
