+++ src/main/java/com/jlubecki/q/logging/LogLevel.java
@@ -1,0 +1,59 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015 Jacob Lubecki
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package com.jlubecki.q.logging;
+
+/**
+ * Determines what should be logged. Illegal states will be logged for every level except {@link #NONE}.
+ */
+public enum LogLevel {
+    /**
+     * Log nothing internally.
+     */
+    NONE,
+
+    /**
+     * Log simple info about the current state.
+     */
+    BASIC,
+
+    /**
+     * Log only items for the Q.
+     */
+    Q,
+
+    /**
+     * Log only items for the players.
+     */
+    PLAYER,
+
+    /**
+     * Log everything to its fullest extent.
+     */
+    FULL;
+
+    public boolean log() {
+        return this != NONE;
+    }
+}
